refactor(game-ambience-sounds): master mixer module now central sound control
